+++ 115211394_Gabriel_Joao_Melissa_Thais/src/cartao/Premium.java
@@ -1,10 +1,18 @@
 package cartao;
 
-
+/**
+ * 
+ * @author Gabriel Alves - Joao Carlos - Melissa Diniz - Thais Nicoly
+ *
+ */
 public class Premium implements TipoDeCartao {
 
+	/**
+	 * metodo que adiciona os pontos ao cartao fidelidade do hospede
+	 * @return o seu bonus
+	 */
 	@Override
-	public int adicionaPontos(float compra) {
+	public int adicionaPontos(double compra) {
 		
 		if (compra > 100){
 			
@@ -26,11 +34,14 @@ public class Premium implements TipoDeCartao {
 		}
 		
 	}
-
+	
+	/**
+	 *  metodo que calcula os descontos das compras do hospede
+	 */
 	@Override
-	public float desconto(float compra) {
+	public double desconto(double compra) {
 
-		float compraComdesconto  = 0;
+		double compraComdesconto  = 0;
 		
 		//desconto de 10% do valor da compra
 		compraComdesconto += compra * 0.9;
